alexa-skill/lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
python-dotenv>=0.15.0
requests==2.26.0